Cellular TLS security change only: fix memory stomp.
When a cellular security context was being free()'ed the profile ID, which is recorded inside the context, was being used AFTER the free() as an index into an array (in order to set a pointer at that index to NULL).  This would likely work most of the time but, on the ESP-IDF platform it just happened to hit something we cared about sufficiently consistently to trace this problem.

Also included is a cosmetic change in the cellSecTlsSettings() test.