STM2F4 platform change only: fix UART flow control setting for "none" case. (#262)
The logic that sets the UART block's flow control register was incorrect: it would have resulted in CTS flow control being selected if the intended flow control was actually none.  We never saw this cause a problem but it is possible we don't have any HW that does not use flow control.  Anyway, this commit fixes the potential problem.