AT client fix: handle rubbish that looks like the start of a valid stop tag. (#268)
A subtle bug, found when testing with SARA-R412M. SARA-R412M sometimes inserts extra carriage returns in the AT response, so for instance:

+BLAH: 1\r\r\n

...rather than the normal:

+BLAH: 1\r\n

The problem with this is that the AT parser sees the "\r", that's fine, it could be part of a valid "\r\n" stop tag, so the AT parser increments its match counter, then it sees another "\r" and so decides this is not actually a "\r\n" stop tag after all and throws it away. Of course, it then doesn't pick up the real "\r\n" stop tag 'cos it's just thrown the initial "\r" from it away. With this commit the AT parser resets the match counter and then checks again to see if a new stop tag has begun.