Automation change for ESP-IDF only: use u_runner when run under automation.
On the ESP-IDF platform we use the native ESP-IDF unit test environment instead of u_runner (since users of ESP-IDF will be used to that) but, with this commit, when run under automation, we switch to using u_runner since that way we can apply filtering on which tests are run and test times can be reduced.  Since the #define U_RUNNER_TOP_STR is always defined by u_run.py we can use it to detect that the build is being done under automation.  We also need to define an environment variable, U_UBXLIB_AUTO, from u_run.py so that the ESP-IDF build system can pick that up and change what is included (otherwise the wrong version of unity_config.h is picked up).

Test: *