Test automation change only: allow multiple filter strings.
When a filter string is passed into u_run_branch.py or u_run.py it can now include multiple filters, separated by a full stop character.  So, for instance, if you want to run the port tests and all of the examples you would use the filter string "port.example"; think of the full stop as an "or".

Test: *