Cellular change only: remove unnecessary power-on delay. (#275)
If the module's VInt line is connected to the MCU and it does not indicate that the module is on then there is no point in checking if the module is alive by bouncing an AT command off it, just go ahead and start powering it up. This saves around 5 seconds in the power-up sequence of a properly wired-up system.