Documentation change only: clarify pin numbering. (#282)
The "pin" numbering used by each SDK is different: for ESP-IDF it is a physical pin number on the ESP32 chip, for STM32F4 and NRF52/NRF53 it is a bit-position in a register bank plus a bank index and has no relation with the physical pin number of the chip; if the chip is inside a u-blox module the pin number of the chip has to be translated to the pin number of the module. This is a [second] attempt to clarify the matter which, particularly in the case of the NINA-Bx modules, can be very confusing since there are three translations to be done: bit/bank position/index -> chip pin -> module pin.