Cellular change: add support for SARA-R422 module.
Note: this commit also introduces a temporary #define, U_CFG_1V8_SIM_WORKAROUND - this is necessary because the SIMs we use in our test farm support 1.8V operation but do not signal that they do so in their answer-to-reset message.  SARA-R422 only supports 1.8V SIMs and, even though these SIMs do work at 1.8V, SARA-R422 by default objects to the lack of 1.8V in the answer-to-reset message and refuses to use them; an AT+UDCONF=92,1,1 command can be applied as a work around.  Once we have sourced some SIMs that do signal 1.8V operation properly the code under this #define will be removed.