Location/GNSS change: add support for Cloud Locate.
Cloud Locate is an assistance service which can be used with a GNSS chip, using the RRLP information from the chip along with assistance information known to u-blox servers to achieve a location fix at relatively low cost.  This commit introduces support for Cloud Locate (which is provided through a subscription to the Thingstream service).

Also included is an implementation of base-64 encoding and decoding; this was used in an intermediate implementation of this commit.  In the end it was not required, however it is included in the commit as a handy utility.

Test automation note: testing of Cloud Locate is enabled only on one test instance, instance 18.  This is because we use a single Thingstream account for Cloud Locate and, since all the tests are run in parallel, enabling more than one could cause random test failures if they happen to overlap.