Cellular change only: add support for AT+UDCONF (#288)
Cellular modules support an AT+UDCONF command which allows details of specific features of the module to be configured, thereafter stored as a non-volatile setting by the module and so this command is generally only needed once in the lifetime of the module. This commit introduces support for setting and reading the most common AT+UDCONF format (i.e. with up to three integer parameters).