zephyr port: Fix stack allocation issue
Zephyr doesn't officially support dynamically allocated stack memory.
But with some hacking it seems that it is possible to get this to work
for at least 32-bit ARM architectures.
This commit will fix the current aligment issues that we sometimes see
for ARM. The stack will now typically be 32 byte aligned for ARM when
CONFIG_MPU_STACK_GUARD is enabled.
There is, however, no guarantee that this will work for other
architectures as alignment requirement may waste a significant amount of
memory. As a minimum detection an assert has been added to check that
aligment is equal or less to 512 bytes.