Gpio implementation cellular (#289)
Cellular change only: add support for GPIOs on the cellular module.

A u-blox cellular module provides GPIO lines, up to six of them, that can be configured as inputs or outputs and then read or set.  This API provides access to those GPIO lines.

IMPORTANT: a cellular module pin may be referred to in three ways:
  - pin number: the physical pin of the cellular module,
  - pin ID: the ID for that pin, which is usually THE SAME AS THE PIN NUMBER,
  - pin name: for instance GPIO1 or SDIO_CMD etc.

This API uses pin ID: do not confuse this with the number on the end of the pin name, i.e. GPIO1 is NOT pin ID 1. Refer to the data sheet for your u-blox module, or section 21.1.2 of the u-blox AT command manual, to determine the pin ID numbers that your particular cellular module supports.

Co-authored-by: Rob Meades <rob.meades@u-blox.com>