zephyr port: Fix thread heap allocation issue
Turns out that there is a bug in Zephyr that results in main threads
reasource pool to be uninitialized. Since the resource pool is copied
from current thread when a new thread is created, all threads will have
uninitialized resource pools unless user assigns a new one using
k_thread_system_pool_assign() or k_thread_heap_assign()..

The result of an uninitialized resource pool is that when a Zephyr API
is used that allocates memory using the thread resource pool, data will
be written to random memory addresses.

As a workaround we assign the system pool to main thread when
uPortInit() is called.