Temporary cellular change only: force AT+UPSV=0 (#302)
On another branch we are integrating support for "UART power saving" (AT+UPSV), where the cellular module is told that it can move to a slow-clocked state and, as a result, not clock the UART, the module being woken up by ubxlib when it is needed. However, some modules store this setting in non-volatile memory and so, the moment we began testing it on the test farm, some of those modules began persistently dropping into this low-power state, even though the control of this state and the corresponding wake-up code has not yet been integrated into master. This led to test failures.

Hence, as a temporary measure, this commit switches this power saving mode off at start of day, so that things return to normal until we have properly tested/reviewed/merged the "UART power saving" commit.