Automation change to ESP-IDF only: remove sending of "*\r" (#305)
The native ESP-IDF unit test environment puts up a menu and requires a test selection to be made, for which *\r\n means "run everything". However, we now use the u_runner test-running mechanism when the code is compiled by the automation scripts and that requires no user input so the sending of *\r\n is redundant; it was originally left in place since it did no harm however u_monitor.py waits for 1 second before sending *\r\n (since the target has to print the menu before it accepts any input) and waiting for 1 second can cause us to lose the initial output from a test if that test begins running immediately, so it actually does cause some harm.

This commit removes it.