Automation change only: improve abort capture (#308)
When the ubxlib test automation system detects an abort on the target it stops the test. However, it does this straight away, cutting off any further logging, when in fact the target can send out useful information after the abort string, e.g. in the ESP-IDF case the abort reason and the back-trace.

This commit delays the end of logging to a couple of seconds after the abort string in order to capture all of the useful information.