BREAKING CHANGE IF YOU HAVE YOUR OWN PORT: add port APIs uPortQueueGetFree() and uPortEventQueueGetFree() (#306)
Two APIs are added to the port API: uPortQueueGetFree(), which returns the number of free entries in the given queue, and uPortEventQueueGetFree(), which does the same for the event queue API (by internally calling uPortQueueGetFree()). If you have your own port then you must implement uPortQueueGetFree(). If it is NOT possible to get the number of free queue entries from your RTOS then simply have uPortQueueGetFree() return U_ERROR_COMMON_NOT_IMPLEMENTED and the porting tests will still pass and the operation of ubxlib will not be adversely affected.