uPortEventQueue change only: fix for potential close() problem.
When an event queue is closed the code did not check that the sending of the `U_EVENT_CONTROL_EXIT_NOW` control message was successful; it could potentially fail if the queue was full and as a result the code would get stuck waiting on the mutex that indicates the event task is running.  With this change the code waits for the send to definitely succeed, retrying until it does, and the note that shut-down of an event queue is a cooperative affair is re-worded somewhat to make it more clear that the user's event callback must compete before a shut-down of the event queue will complete.