Platform change: add support for Arduino, on ESP32 only.
At customer request, this commit adds support for building ubxlib inside Arduino, for the ESP32 platform only, allowing the ubxlib C APIs to be called by customers who normally write to the Arduino C++ peripheral APIs.

Full instructions are provided in the README.md file /port/platform/arduino/README.md but basically there is a Python script which arranges ubxlib into a form that the Arduino build tools can understand and an example sketch which uses the ubxlib APIs to open a socket to an echo server and exchange some data.  Other than that all of the ubxlib APIs are available: they are not "Arduinoised" or "C++-ised" in any way, they are simply made available to Arduino users; the ubxlib C-code examples should make it obvious what other things can be done.  The usual Arduino APIs can, of course, be called as well.

The test automation is updated so that the ubxlib tests, as well as the example sketch, are built and tested under Arduino (on a NINA-W1 EVK and an ESP32 DevKitC board) in the same way as any other platform.