Test change for ESP-IDF platform only: correct OS event queue task priority.
In the porting test of the event queues, the priority of the tasks at the end of the event queues must be higher than that of the task sending messages to them, otherwise we can't be sure that they will complete their work in time.  This wasn't the case for ESP-IDF; they were running at the same priority as the task sending messages to them, and hence the event queue tests could sometimes fail.  Also, one of the assert checks was on the wrong counter value, which is now fixed.