Cellular SARA-R5 module change only: boot time and radio-off style. (#324)
The actual SARA-R5 boot time, measured by the appearance of the greeting message, is closer to 6 seconds than 5.  The radio-off command for SARA-R5 was set to be AT+CFUN=0 but AT+CFUN=0 doesn't persists across power cycles, leading to the possibility that the module will start to register after boot at about the point we are setting the AT+CFUN mode, which is not desirable, hence with this commit we switch to using AT+CFUN=4 (airplane mode) for radio power off as that _does_ persist across power cycles.  This entails a slight modification to the RAT setting commands since SARA-R5 only permits the RAT to be changed when it is in AT+CFUN=0, not AT+CFUN=4.