Cellular change only: add API to set auto-bauding on/off.
Where supported, it is usual to leave the cellular module as auto-bauding, where the module will detect the baud rate of the first characters incoming from the MCU UART after the module has booted and will adopt that baud rate.  However, if one wants to have a greeting message emitted by the module the moment the module has booted (rather than having to wait for the first few characters to arrive at the module) then an API is introduce to disable auto-bauding, fixing the baud rate at the current UART rate.

Note that if you run the ubxlib cellular configuration tests they will switch auto-bauding off and that setting will be stored in the module's non-volatile memory.  This so that we can set a greeting message and hence tell exactly when the module has booted.  For modules that support auto-bauding (e.g. SARA-R5 and SARA-U201), if you want auto-bauding back on again after doing this, your application must initially use 115200 as the UART baud rate (the rate that the tests will set it to) and then call uCellCfgSetAutoBaudOn() (and ensure that uCellPwrOff() is used to power the cellular module off afterwards).