Cellular sockets change only: add ability to use hex mode. (#309)
Currently the cellular sockets API, at AT interface-level, sends and receives binary-format data between the MCU and the cellular module. While this is apparently optimal it involves telling the cellular module how much data is to be sent for each packet, waiting for a prompt, then actually sending the data. For short packets this exchange can take between 50 and 100 ms, longer than it would take to just send the whole packet immediately (in hex mode), even though this involves sending and receiving twice as many bytes at the interface.

This commit adds three APIs in the cellular sockets API: uCellSockHexModeOn(), uCellSockHexModeOff() and uCellSockHexModeIsOn(), allowing the user to chose hex mode for more efficient operation when sending smallish (e.g. 50 bytes) chunks of data in rapid succession. Note that the maximum datagram size is halved (i.e. reduced to 512 bytes) in this mode and the mode applies to all of the sockets in use on the cellular module (i.e. it is not "per socket").