short range edm: Fix stream stall
In cases where an EDM event doesn't end up in the event queue the EDM
stream could end up getting stuck. This can, for instance, happen when a
disconnect event is received for an unknown peer. In this case the event
should marked as processed so next EDM event can be handled.