stm32 port: Add workaround for newlib memory leak
When _LITE_EXIT is enabled the stdio streams stdout, stdin and stderr are not closed
when deallocating the task, resulting in memory leaks if the deleted task have been using
these io streams.

As a workaround we manually close the stdio streams when user calls
uPortTaskDelete().

Note: The workaround only works when a task deletes itself.

Fixes #310