Documentation change only: clarify uPortQueueGetFree() is optional.
Add a note to the uPortQueueGetFree() function documentation to indicate that it is not a requirement that the function is implemented, it can return U_ERROR_COMMON_NOT_IMPLEMENTED and the operation of ubxlib will not be adversely affects.  This note was already present on uPortEventQueueGetFree() but was missed here.