BREAKING CHANGE IF YOU HAVE YOUR OWN PORT: add uPortQueuePeek().
uPortQueuePeek() is added to the port API. This functions allows the message that would be uPortQueueReceived() to be read WITHOUT being removed from the message queue.  If you have your own port then you must implement uPortQueuePeek(). If it is NOT possible to implement this with your RTOS then simply have uPortQueuePeek() return U_ERROR_COMMON_NOT_IMPLEMENTED; the porting tests will still pass and the operation of ubxlib will not be adversely affected.