test: Add workaround for newlib FILE allocations
When newlib is build with _REENT_GLOBAL_STDIO_STREAMS *disabled*
a global dynamic pool will be used for FILE pointers.
The pool re-uses existing FILE pointers but if no FILE is currently
free for use, newlib will allocate a new one. Since our tests checks
heap usage this can result in "false" memory leak failurs.
To mitigate this problem we start with allocating a couple of FILE
pointers so that that newlib doesn't need to allocate any new ones
throughout the complete test suite.

TODO: REMOVE THIS WHEN #272 IS DON