event_queue: Add workaround for newlib delayed stdout allocation
This is a temporary workaround to prevent false memory leak failures
in our automated tests.

When _REENT_SMALL is enabled in newlib the allocation of the stdout
stream is delayed until it is needed. To prevent the delayed allocation
we just make an empty print.

TODO: REMOVE THIS WHEN #272 IS DONE