Automation change only: override U_CFG_APP_PIN_CELL_PWR_ON. (#347)
On moving from NCS v1.4.2 to NCS v1.6.1 we found that, on the Nordic NRF5340 DK boards (nrf5340dk_nrf5340_cpuapp) that we use in the test farm, pin 33 (AKA 1.01) that we used to power-on the cellular module no longer worked.  Hence here we override U_CFG_APP_PIN_CELL_PWR_ON to be pin 36 (AKA 1.04).  For backwards-compatibility the definition of U_CFG_APP_PIN_CELL_PWR_ON in u_cfg_app_platform_specific.h is unchanged; instead a note is added there and the pin number is overridden in DATABASE.md.