Fixes in uPortEventQueueSend() and uAtClientCallback(). (#346)
uPortEventQueueSend(), as with all of the other uPortEventQueue* functions, locks the mutex that protects the linked-list on entry and unlocks it on exit. However, in the middle of this function uPortQueueSend() is called and that might block if the queue it is sending to is full, and, with the mutex that protects the linked-list being locked, this leads to the entire uPortEventQueue API being locked for that time. This is unnecessary, the mutex only needs to protect the linked list, hence with this commit the lock is released before uPortQueueSend() is called. A consequence of this is that an event queue should not be closed while uPortEventQueueSend() is in progress, which is made clear in the function documentation.

uAtClientCallback() performs a callback using an event queue that is created within uAtClient. The priority of the task at the end of that event queue used to be set very low (U_CFG_OS_PRIORITY_MIN + 2); however uAtClientCallback() is often used in a chain of event tasks, which are customarily set to U_CFG_OS_APP_TASK_PRIORITY, and having such a low priority in the chain leads to a "kink", with the queue that terminates at the uAtClient callback task filling up 'cos it has not been scheduled. Coupled with the uPortEventQueueSend() issue above this may even lead to character loss in the UART driver, if that was the source of events, since uPortQueueSend() may be unable to run anywhere in the chain of event queues (because of the linked-list mutex issue) and uPortQueueSendIrq() would fail as the queue it is sending to becomes full. With this commit the task at the end of the uAtClientCallback() queue is set to run at U_CFG_OS_APP_TASK_PRIORITY in order to be quite sure it doesn't hold anything up.