Security API change: add ability to use device certificate. (#350)
IMPORTANT: this change adds two new members to the uSecurityTlsSettings_t structure (useDeviceCertificate and includeCaCertificates); U_SECURITY_TLS_SETTINGS_DEFAULT is modified to match (setting both values to `false`).  Hence, provided you have followed the rule that any variable of type uSecurityTlsSettings_t is assigned the values of U_SECURITY_TLS_SETTINGS_DEFAULT when it is created, no modifications to your code are required as a result of this commit.

When a [currently only cellular] module that supports u-blox security has been security sealed, an X.509 public device certificate and private key pair are created and stored in the module.  With this commit the X.509 public device certificate and private key pair generated during the sealing process, and the CA X.509 certificates used to sign them, can be read and can also be "implicitly" used with TLS security without the need to read them.  Note that:

- the ability to read the certificates will only work if Zero Touch Provisioning is enabled on the relevant device profile in your Thingstream account,
- the ability to "implicitly" use the certificates with TLS security without reading them first is not supported on all of the modules that support security sealing (if you see what I mean); check the AT manual for your module, the AT+USECPRF=14 command, for further information; if opcode 14 is not included for AT+USECPRF, or is present but not indicated as supported for your module type, then this feature is not supported by your module.