Cellular file system change only: add ability to use tags. (#353)
The cellular module file system supports read/write/delete/list of files from pre-defined "tagged" areas of the file system, a little like directories but the tags are ONLY pre-defined by the module. Call uCellFileSetTag() to set the current tag and pUCellFileGetTag() to read the currently active tag. Note that uCellFileBlockRead() does NOT support use of tags; this is a limitation of all the cellular modules currently supported by ubxlib.