Cellular change only: add APIs to get/set serial interface profile (#354)
Functions are added to the cellular API to configure the requested variant of serial profile or get the currently active profile, allowing the interpretation of data over the serial interface to be modified to carry, for instance, AT modem, digital audio or trace data.

The configuration supported by each cellular module type is generally quite specific to that module type.  Note that no tests of this feature are included since the serial profiles employed by the modules running on the ubxlib test farm is necessarily fixed; instead this commit was tested and confirmed to work manually.