Temporary security [cellular] test change only: remove security heartbeat call. (#356)
One of the test instances (instance 13) has begun misbehaving in that it takes a very long time to return an answer to AT+USECCONN, i.e. the uSecurityHeartbeatTrigger()/uCellSecHeartbeatTrigger() functions.  With this commit the call to uSecurityHeartbeatTrigger() is temporarily removed from the securityE2eEncryption and securityPskGeneration tests while the cause is being investigated.