nrf5 port: Move out parts from Makefile to a shared ubxlib.mk
This will allow us to have a shared ubxlib.mk Makefile that can be
re-used by ports which are Makefile based.

ubxlib.mk works in the same way as the shared CMake ubxlib.cmake file.
I.e. it is used to get a list of the ubxlib source code files and
include directories. Details of how to use ubxlib.mk can be found in
port/platform/README.md.