doc: Add note about STM32CubeIDE project being replaced with Makefile (#369)
Heads up for our users that we are about to remove the STM32CubeIDE
project from the ubxlib repo.

The message is embedded as HTML in readme_images/important_msg.svg.
GitHub have very restricted inline HTML support so there seems to be no
way of changing background color of things. But by using a .svg file and
the "foreignObject" it is possible.