zephyr port: Make ubxlib a zephyr module (#365)
To make integration of ubxlib easier for Zephyr users this commit adds
what is needed to make ubxlib a Zephyr module. The module includes
Kconfig that will let the user configure ubxlib using Zephyr menuconfig.
In this way the user doesn't need to know what Zephyr features must be
enabled for a certain ubxlib feature.

This commit also includes a shared ubxlib.cmake file which is used to
collect all source files and include directories. This is currently
shared between the Zephyr and IDF port in order to minimize the number
of places all ubxlib source files are listed.

To use the Zephyr module one can either append:
   -DZEPHYR_EXTRA_MODULES=<ubxlib_path>
to the west command OR specify it in the application CMakeLists.txt by
adding the following line at the top of the file:
   list(APPEND ZEPHYR_EXTRA_MODULES ${UBXLIB_BASE})
Note: When adding it to west command line it will always recompile.