New platform: Windows (win32).
This commit adds support for Windows (win32) as a ubxlib platform. The Windows platform is intended primarily for debug/development use rather than as a product platform and beware that it will likely change slightly over the next few months as we settle it down. CMake is used, which means that the code can be very conveniently built and debugged in Visual Studio Code with the aid of the CMake Tools extension; this is further described in the README.md in the runner build directory. Compilation is under MSVC.

Some minor additional changes are made in the test code to permit compilation under MSVC.

To use a real u-blox module with the platform, simply plug a u-blox EVK into the USB port of a PC and use the number of that COM port as your UART number with the ubxlib APIs.

Note that, while Windows is a great platform for rapid application development and visibility, heap and stack checking are not performed under Windows.

Test automation note: a new test instance is added to run the Windows platform but only local tests, i.e. ones that do not require an attached module or a UART loop-back, will currently run since there is no u-blox module connected to the test machine for this purpose; this will be fixed early in the new year.

IMPORTANT: if you are running the ubxlib test automation scripts locally then they will STOP after this commit since there are new machine-specific entries associated with the new Windows platform that MUST be populated. To resolve this:

- if you are running on Windows, install MSVC by going to https://aka.ms/vs/17/release/vs_BuildTools.exe and, in the installation tool which runs, selecting to install "Desktop development with C++",
- open the file C:\Users\<your user name>\.ubx_automation\settings_v2_agent_specific.json in a text editor and search for the string "_FIX_ME": remove the "_FIX_ME" post-fixes from the ends of the string entries it finds, and modify their values to match the version of the MSVC you downloaded, if different, and save the file.

If you are NOT running on Windows then don't try to run test instance 23 as it is Windows only.