vscode: Refactor scripts and provide better package management
The installation of dependencies have now been isolated to a separate
module called u_packages. This is a very simple package manager that
uses .vscode/u_packages.yml as input to what packages to install.

Each time you start a pyinvoke task (such as "inv nrf5.build") the
task will first check if all the packages needed for this specific task
are installed. If they are not the user will be prompted whether to
automatically install them or not.

All packages will be installed in the ${HOME}/.ubxlibpkg by default.
You can override this by using the UBXLIB_PKG_DIR env. variable.

Each package has a version number and each version of a package is
stored in a separate folder. When you switch version in u_packages.yml
the package manager will first check if the specific version can be
found in the local package dir and if not the user will be prompted
whether to automatically install it.