Cloud Locate fix only
The internal stringToInt32() function which parses the response from Cloud Locate was considering it an error if a measurement didn't have the full number of digits, e.g. a latitude of 52.2227347 was OK but a latitude of 52.222734 (i.e. only 6 digits in the fractional part) was not. This commit fixes that.