Arduino example change only: better error messages. (#395)
The Arduino example code doesn't print a clear error message if uNetworkAdd() fails, it just continues and fails a bit later. This commit fixes things so that it prints an error message at that point and the code then goes on to do nothing else, making the location of the error plain.