Cellular power-related change only: allow enable power pin sense to be inverted. (#396)
The sense of pinEnablePower passed into uCellAdd() is that a 1 on that pin means power is enabled and a 0 means it is not.  However, it is possible that a customer's HW has this the other way up, hence this commit introduces a #define U_CELL_ENABLE_POWER_PIN_INVERTED which, when defined, inverts the sense such that a 0 on pinEnablePower means power is enabled and a 1 means it is not.