BREAKING CHANGE IF YOU HAVE YOUR OWN PORT: add uPortUartCtsSuspend()/uPortUartCtsResume()
This commit adds an API to the UART porting layer which permits CTS (i.e. towards the MCU) flow control to be suspended and resumed.  This is required where the [cellular] module at the other end of the UART can enter a sleep state during which the CTS line may be de-asserted, preventing characters being sent to the module, which is no good if those characters are the stimulus that causes the module to awaken from its sleep state.

This commit is JUST the API, the header file; the implementation and tests are added in subsequent commits.

This API does NOT HAVE TO BE SUPPORTED by a given platform; where it is not supported U_ERROR_COMMON_NOT_SUPPORTED can be returned.  Obviously the ability to put the device at the other end of the UART to sleep may be compromised where this is the case.