Cellular change only: UART power saving when HW flow control is enabled.
Use of HW flow control with a cellular module complicates matters for UART (AT+UPSV) power-saving since on some modules (e.g. SARA-R5) the CTS pin of the module floats high while it is asleep, preventing the wake-up character being sent to the module to get it out of UART sleep.  With this commit the new uPortUartCtsSuspend()/uPortUartCtsResume() functionality is used to work around this problem for all platforms except Zephyr.

Zephyr is problematic in that the use of HW flow control is a compile-time setting, it cannot be changed at run-time.  Hence, for this case, the "DTR power saving" feature of the SARA-R5 module is now also supported; an interface is added to the AT client (uAtClientSetActivityPin()) that can set a pin which will be asserted while the AT client is communicating with the target and de-asserted when it stops communicating with the target.  The function uCellPwrSetDtrPowerSavingPin() allows the DTR pin to be set in the cellular API and the cellular API will pass that pin to the AT client as soon as it has configured UART power saving in the cellular module.

Since the DTR pin is not included in the configuration structure that is passed to the network API, a conditional compilation flag
U_CFG_APP_PIN_CELL_DTR can be used for this case; this is an exception to the usual rule that all configuration items are passed-in as run-time parameters, but that is fine for Zephyr since all the Zephyr UART peripheral pins are compile-time defined in any case.

Note that the DTR mechanism does not apply to the SARA-R4xx modules since they do not support DTR wake-up; they hold their CTS pin low while asleep and so no problem arises, a wake-up character can always be sent.