pytasks: Add terminal command to esp_idf and nrfconnect
The terminal command will open a development terminal where the shell
environment is setup for the associated SDK.

For nrfconnect (Zephyr) this means that ZEPHYR_BASE will be set so that
the west command can be used for building and flashing targets.

For esp_idf the ESP-IDF export.sh will be loaded so that idf.py can be
used.