Windows platform change: treat char as unsigned.
In Microsoft Visual C++ char types are signed, which can lead to unexpected behaviours e.g. a character value which contains 0xaa, when compared with the literal value 0xaa, will return false; the character value is interpreted as being negative because it has the top bit set, while the literal value 0xaa is positive.  To avoid this problem advice is added that the command-line switch /J should be used with the compiler to force char types to be treated as unsigned.

A check is added to u_port.c that _CHAR_UNSIGNED is defined for MSVC (/J just defines that compilation flag, which has an effect in limits.h) and the CMakeLists.txt for the runner build is updated to define it.

Then all the nasty casts can go away.