MQTT fix for cellular: permitted combination of keep alive and inactivity timeout. (#412)
MQTT fix for cellular: permitted combination of keep alive and inactivity timeout.

The values of the MQTT inactivity timeout and MQTT keep-alive are implicitly linked: keep-alive is irrelevant if there is no inactivity timeout, since the MQTT "ping" packet is only sent when the inactivity timer is about to expire.  However, there is an additional complication in that, for SARA-R5 cellular modules, the keep-alive functionality interprets an inactivity timeout of 0 to mean "expire immediately" and so will send MQTT ping packets constantly, usually resulting in a "queue full" memory error for all subsequent MQTT operations.

With this commit a cross-check is added so that keep-alive cannot be turned on if the inactivity timeout is zero.