ESP32 change only: increase stack of C2C test task. (#411)
On an Arduino build (instance 11.1) the test task used in the C2C test (cellSecC2cAtClient()) was throwing a "stack canary" exception. After adding 1024 bytes to the 2048-byte stack of that task and running again, the ESP-IDF stack extent checking reckoned we would have been 56 bytes inside the original stack limit (i.e. we had a minimum of 1080 bytes free) but, of course, increasing the stack did fix the problem.

It is difficult to know whether there is an issue elsewhere or whether the stack extent checking is not 100% accurate. This commit increases the stack size for this test task by 256 bytes to 2304 bytes; not much else I can think of to do right now.