Add U_ASSERT() macro and assert hook registration function. (#418)
Until now ubxlib has called the C-library assert() macro in various places.  With this commit all such references are removed and instead a U_ASSERT() macro is defined; should the condition passed to it be false, calls uAssertFailed() will be called with the file name and line number of U_ASSERT().

The user can override uAssertFailed(), either using the linker (since uAssertFailed() is weakly linked) or by calling uAssertHookSet() to set an alternative assert failure function.