Comment change to clarify need for U_CFG_SARA_R5_M8_WORKAROUND. (#428)
Update the comment concerning U_CFG_SARA_R5_M8_WORKAROUND to identify when it needs to be applied, i.e. only for prototype versions of the SARA-R510M8S module, production week (printed on the upper right of the module label) less than 20/27; so not at all really, except on some of the very old module HW we use in the ubxlib test farm.