Cellular network fix: potential array overrun during network search. (#427)
In uCellNetScanGetFirst(), uAtClientReadBytes() is used when reading the list of networks because the text it is reading contains quotation marks, which uAtClientReadString() would remove; however that means that we don't end up with a terminator on the end, so the strtok_r() calls which follow may wander off into space; a terminator is now added explicitly in uCellNetScanGetFirst(). This was found when testing on Windows as Windows is able to detect such things.